scripts/requirements.txt
requests
beautifulsoup4
python-dotenv
supabase-client
